hotfix: data coversion in inference

+++ mmv_im2im/proj_tester.py
@@ -68,7 +68,7 @@ class ProjectTester(object):
             x = np.expand_dims(x, axis=0)
 
             # TODO: add convert to tensor with proper type, similar to torchio check
-            x = torch.tensor(x).cuda()
+            x = torch.tensor(x).float().cuda()
 
             if (
                 self.model_cfg.model_extra is not None
